+++ packages/terra-core-docs/src/terra-dev-site/test/toggle-button/IconAnimated.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ToggleButton from 'terra-toggle-button';
+
+const AnimatedIconToggleButton = () => (
+  <ToggleButton id="animatedIcon" closedButtonText="ToggleButton" isIconAnimated>
+    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+  </ToggleButton>
+);
+
+export default AnimatedIconToggleButton;