+++ packages/terra-core-docs/src/terra-dev-site/test/toggle-button/IconOnly.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ToggleButton from 'terra-toggle-button';
+
+const IconOnlyToggleButton = () => (
+  <ToggleButton id="iconOnlyToggleButton" closedButtonText="Custom Text" isIconOnly>
+    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+  </ToggleButton>
+);
+
+export default IconOnlyToggleButton;