+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Added
   * Migrate docs from terra-switch.
+  * Migrate docs from terra-toggle-button.
+  * Migrate docs from terra-visually-hidden-text.
+  * Migrate docs from terra-toolbar.
+  * Migrate docs from terra-text.
+  * Migrate docs from terra-toggle-section-header.
 
 ## 1.12.2 - (January 4, 2022)