+++ packages/terra-toggle-section-header/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs to terra-core-docs.
+
 ## 2.57.1 - (January 4, 2022)
 
 * Changed